requirements.txt
streamlit
pandas
openai==0.28
langchain 
python-dotenv 
sentence-transformers
